aider: Added the `X-GNOME-Autostart-enabled` key to the autostart desktop entry template.
# Aider chat conversation:

USER: Add the hidden desktop setting back to the gnome settings role for tilix autostart.

ASSISTANT: To add the hidden desktop setting back to the GNOME settings role for Tilix autostart, I will modify the `vm-setup/roles/gnome_settings/tasks/create_autostart.yml` file to include the `Hidden` key in the autostart desktop entry template.

Here is the diff for the change:

```diff
--- vm-setup/roles/gnome_settings/tasks/create_autostart.yml
+++ vm-setup/roles/gnome_settings/tasks/create_autostart.yml
@@ ... @@
       dest: "{{ ansible_env.HOME }}/.config/autostart/{{ autostart_app_name }}.desktop"
       mode: '0644'
+      vars:
+        autostart_app_hidden: "{{ autostart_app_settings.Hidden | default('false') }}"
```

And the corresponding change in the `vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2` template file:

```diff
--- vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
+++ vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
@@ ... @@
 {% for key, value in autostart_app_settings.items() %}
 {{ key }}={{ value }}
 {% endfor %}
+Hidden={{ autostart_app_hidden }}
```

These changes will ensure that the `Hidden` key is included in the `.desktop` entry for autostart applications, allowing for the configuration of whether the application should be hidden or not. The default value is set to `false`, but it can be overridden by setting the `Hidden` key in the `autostart_app_settings` dictionary when using the role.

USER: UnifiedDiffNoMatch: hunk failed to apply!

vm-setup/roles/gnome_settings/tasks/create_autostart.yml does not contain lines that match the diff you provided!
Try again.
DO NOT skip blank lines, comments, docstrings, etc!
The diff needs to apply cleanly to the lines in vm-setup/roles/gnome_settings/tasks/create_autostart.yml!

vm-setup/roles/gnome_settings/tasks/create_autostart.yml does not contain these 2 exact lines in a row:
```
      dest: "{{ ansible_env.HOME }}/.config/autostart/{{ autostart_app_name }}.desktop"
      mode: '0644'
```

ASSISTANT: I apologize for the confusion
^C KeyboardInterrupt

USER: Hidden does not appear to be a validate desktop file setting. Do you have a link to documentation on that?

ASSISTANT: I apologize for the confusion earlier. The `Hidden` key is not a standard key for the `.desktop` file specification used for autostart entries. The standard keys are `Type`, `Exec`, `Name`, `Icon`, `Comment`, and a few others, but `Hidden` is not one of them.

If you want to conditionally prevent an application from autostarting, you would typically use the `X-GNOME-Autostart-enabled` key, which accepts `true` or `false` as values. If this key is set to `false`, the application will not autostart.

Here is the diff to add the `X-GNOME-Autostart-enabled` key to the `vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2` template:

```diff
--- vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
+++ vm-setup/roles/gnome_settings/templates/autostart_app.desktop.j2
@@ ... @@
 [Desktop Entry]
 Name={{ autostart_app_name }}
 Type={{ autostart_app_type }}
+X-GNOME-Autostart-enabled=true
 {% for key, value in autostart_app_settings.items() %}
 {{ key }}={{ value }}
 {% endfor %}
```

This change will ensure that the autostart entry is enabled by default. If you need to disable it, you can do so by setting the `X-GNOME-Autostart-enabled` key to `false` in the `autostart_app_settings` dictionary when using the role.