aider: Moved the Docker group from `install_pkgs` to `docker` role's `vars/main.yml` and renamed it to `docker_packages`.
# Aider chat conversation:

USER: Move the docker group in install packages to the docker role var/main.yml and name it docker_packages.

ASSISTANT: To move the Docker group from the `install_pkgs` role to the `docker` role's `vars/main.yml` and rename it to `docker_packages`, I will make the following changes:

1. Remove the Docker group from `vm-setup/roles/install_pkgs/defaults/main.yml`.
2. Add the Docker group to `vm-setup/roles/docker/vars/main.yml` and rename it to `docker_packages`.

Here are the diffs for those changes:

```diff
--- vm-setup/roles/install_pkgs/defaults/main.yml
+++ vm-setup/roles/install_pkgs/defaults/main.yml
@@ ... @@
     - unzip
     - vim
     - wget
     - zip
   gui:
     - fonts-inconsolata
     - gnome-tweaks
     - tilix
-  docker:
-    - docker-ce
-    - docker-ce-cli
-    - containerd.io
-    - docker-compose-plugin
```

```diff
--- /dev/null
+++ vm-setup/roles/docker/vars/main.yml
@@ ... @@
+docker_packages:
+  - docker-ce
+  - docker-ce-cli
+  - containerd.io
+  - docker-compose-plugin
```

These changes will move the Docker package list from the `install_pkgs` role to the `docker` role and rename the variable to `docker_packages`.