aider: Convert the installation of prerequisites to use a meta task with the `install_pkgs` role.
# Aider chat conversation:

USER: Conver the instal prereq's to use a meta task using the install packages role

ASSISTANT: To convert the installation of prerequisites to use a meta task with the `install_pkgs` role, we need to modify the `vm-setup/roles/nodejs/tasks/main.yml` file. We will remove the direct `apt` task and instead include the `install_pkgs` role with the necessary packages as variables.

Here is the diff for the change:

```diff
--- vm-setup/roles/nodejs/tasks/main.yml
+++ vm-setup/roles/nodejs/tasks/main.yml
@@ ... @@
-- name: Install prerequisites for nvm
-  become: yes
-  apt:
-    name:
-      - curl
-      - build-essential
-    state: present
-    update_cache: yes
+
+- name: Include install_pkgs to install prerequisites for nvm
+  include_role:
+    name: install_pkgs
+  vars:
+    packages:
+      - curl
+      - build-essential
```

This change will make the `nodejs` role use the `install_pkgs` role to install the prerequisites for `nvm`, adhering to the DRY (Don't Repeat Yourself) principle and reusing the existing role for package installation.