aider: Move the call to install packages to roles/nodejs/meta/main.yml.
# Aider chat conversation:

USER: move the call to install packages to roles/nodejs/meta/main.yml

ASSISTANT: To move the call to install packages from the `vm-setup/roles/nodejs/tasks/main.yml` file to the `vm-setup/roles/nodejs/meta/main.yml` file, we will create a new `meta/main.yml` file for the `nodejs` role with the necessary dependencies.

Here is the diff for the new `meta/main.yml` file in the `nodejs` role:

```diff
--- /dev/null
+++ vm-setup/roles/nodejs/meta/main.yml
@@ ... @@
+---
+dependencies:
+  - role: install_pkgs
+    vars:
+      packages:
+        - curl
+        - build-essential
```

And here is the diff for the changes to the `vm-setup/roles/nodejs/tasks/main.yml` file to remove the now redundant task:

```diff
--- vm-setup/roles/nodejs/tasks/main.yml
+++ vm-setup/roles/nodejs/tasks/main.yml
@@ ... @@
-- name: Include install_pkgs to install prerequisites for nvm
-  include_role:
-    name: install_pkgs
-  vars:
-    packages:
-      - curl
-      - build-essential
```

These changes will move the installation of prerequisites for `nvm` into the `meta/main.yml` file of the `nodejs` role, making it a dependency that is handled before the role's tasks are executed.