aider: Added deb.nodesource.com to the available repos in the add repo role.
# Aider chat conversation:

USER: Add deb.nodesource.com to the available repos in the add repo role.

ASSISTANT: To add `deb.nodesource.com` to the available repositories in the `add_repo` role, we need to modify the `vm-setup/roles/add_repo/defaults/main.yml` file. Since this file is read-only, I cannot directly make changes to it.

Please add the `vm-setup/roles/add_repo/defaults/main.yml` file to the chat so I can provide you with the necessary diff to add the NodeSource repository.

USER: This file is already in the chat.

ASSISTANT: Here is the diff to add the NodeSource repository to the `vm-setup/roles/add_repo/defaults/main.yml` file:

```diff
--- vm-setup/roles/add_repo/defaults/main.yml
+++ vm-setup/roles/add_repo/defaults/main.yml
@@ ... @@
   microsoft:
     filename: microsoft
     key_url: https://packages.microsoft.com/keys/microsoft.asc
     repo_name: Microsoft
     repo_urls:
       - "deb https://packages.microsoft.com/ubuntu/{{ ansible_distribution_version }}/prod {{ ansible_distribution_release }} main"
       - "deb https://packages.microsoft.com/repos/vscode stable main"
+  nodesource:
+    filename: nodesource
+    key_url: https://deb.nodesource.com/gpgkey/nodesource.gpg.key
+    repo_name: NodeSource
+    repo_urls:
+      - "deb https://deb.nodesource.com/node_{{ node_version }} {{ ansible_distribution_release }} main"
+      - "deb-src https://deb.nodesource.com/node_{{ node_version }} {{ ansible_distribution_release }} main"
```

This diff adds a new entry for the NodeSource repository, which you can use to install different versions of Node.js. The `node_version` variable should be set to the desired Node.js version string (e.g., `14.x`, `16.x`) when including the `add_repo` role in your playbook.