aider: Hardcoded the `node_version` variable to `20.x` in the `add_repo` role.
# Aider chat conversation:

USER: Change the variable node_version to a hardcoded 20.x

ASSISTANT: To hardcode the `node_version` variable to `20.x` in the `add_repo` role, we need to modify the `vm-setup/roles/add_repo/defaults/main.yml` file. Please add this file to the chat so I can provide you with the necessary diff.

USER: that file is already in the chat

ASSISTANT: Here is the diff to hardcode the `node_version` variable to `20.x` in the `vm-setup/roles/add_repo/defaults/main.yml` file:

```diff
--- vm-setup/roles/add_repo/defaults/main.yml
+++ vm-setup/roles/add_repo/defaults/main.yml
@@ ... @@
   nodesource:
     filename: nodesource
     key_url: https://deb.nodesource.com/gpgkey/nodesource.gpg.key
     repo_name: NodeSource
     repo_urls:
-      - "deb https://deb.nodesource.com/node_{{ node_version }} {{ ansible_distribution_release }} main"
-      - "deb-src https://deb.nodesource.com/node_{{ node_version }} {{ ansible_distribution_release }} main"
+      - "deb https://deb.nodesource.com/node_20.x {{ ansible_distribution_release }} main"
+      - "deb-src https://deb.nodesource.com/node_20.x {{ ansible_distribution_release }} main"
```

This change replaces the variable `{{ node_version }}` with the hardcoded version string `20.x` for the NodeSource repository URLs.