aider: Add version tracking functionality to the `github_release` role.
# Aider chat conversation:

USER: Is there a way in ansible to maintain a local file with version information. E.g., for the github_release role, if this the first time it's used it saves something like 'name of file => installed version' to a file and when a new version is downloaded that line is updated. This would be one file for all files github_release was used for.

ASSISTANT: Yes, Ansible can be used to maintain a local file with version information. To achieve this, you can use the `lineinfile` module to ensure that a specific line is present in a file. This line would contain the name of the file and the installed version. When a new version is downloaded, you can update the line in the file with the new version information.

Here's how you can modify the `github_release` role to maintain a local version tracking file:

1. Add a task to read the current version from the version tracking file (if it exists) and register a variable with that information.
2. Add a task to download the new version if there's a version mismatch.
3. After a new version is downloaded, add a task to update the version tracking file with the new version information.

Let's create a new file called `version_tracking.yml` in the `tasks` directory of the `github_release` role to encapsulate the version tracking logic. Here's the diff for creating this new file:

```diff
--- /dev/null
+++ vm-setup/roles/github_release/tasks/version_tracking.yml
@@ ... @@
+# This file contains tasks for tracking the version of downloaded assets
+
+- name: Check if version tracking file exists
+  stat:
+    path: "{{ github_version_tracking_file }}"
+  register: version_tracking_file
+
+- name: Read current version of the asset from the version tracking file
+  command: "grep '{{ github_asset_name }}' {{ github_version_tracking_file }} | cut -d '=' -f 2"
+  register: current_asset_version
+  changed_when: false
+  failed_when: false
+  when: version_tracking_file.stat.exists
+
+- name: Update version tracking file with the new version
+  lineinfile:
+    path: "{{ github_version_tracking_file }}"
+    regexp: '^{{ github_asset_name }}='
+    line: "{{ github_asset_name }}={{ github_latest_version }}"
+    create: yes
+  when: github_version_mismatch | default(true) and github_download_result is changed
```

Next, we need to include this new task file in the main task file after the download task. Here's the diff for the `tasks/main.yml` file:

```diff
--- vm-setup/roles/github_release/tasks/main.yml
+++ vm-setup/roles/github_release/tasks/main.yml
@@ ... @@
 - name: Download file
   get_url:
     url: "{{ github_download_url }}"
     dest: "{{ github_dest_path }}"
     mode: '{{ github_file_mode | default("0644") }}'
   when: github_version_mismatch | default(true)
   register: github_download_result

+- name: Include version tracking tasks
+  include_tasks: version_tracking.yml
+  vars:
+    github_asset_name: "{{ github_asset_name_pattern }}"
+    github_version_tracking_file: "/path/to/version_tracking_file"
+
 - name: Ensure the downloaded file has the proper permissions
   file:
     path: "{{ github_dest_path }}"
     mode: '{{ github_file_mode | default("0644") }}'
     state: file
   when: github_version_mismatch | default(true) and github_download_result is changed
```

Make sure to replace `/path/to/version_tracking_file` with the actual path where you want to store the version tracking information.

This implementation assumes that the `github_asset_name_pattern` variable can be used to uniquely identify the asset within the version tracking file. If this is not the case, you may need to adjust the logic to suit your needs.