aider: Updated version tracking file path to be relative to the playbook directory and hardcoded as `.github_release_installed_versions`.
# Aider chat conversation:

USER: This tracking file path would be the location of the playbook being used and the filename would be hardcoded to .github_release_installed_versions

ASSISTANT: To implement the requested change, I will update the `github_release` role's `version_tracking.yml` file to use a hardcoded path for the version tracking file, which will be relative to the playbook's directory and named `.github_release_installed_versions`.

Here is the diff for the change:

```diff
--- vm-setup/roles/github_release/tasks/version_tracking.yml
+++ vm-setup/roles/github_release/tasks/version_tracking.yml
@@ ... @@
-    path: "{{ github_version_tracking_file }}"
+    path: "{{ playbook_dir }}/.github_release_installed_versions"
   register: version_tracking_file

@@ ... @@
-  command: "grep '{{ github_asset_name }}' {{ github_version_tracking_file }} | cut -d '=' -f 2"
+  command: "grep '{{ github_asset_name }}' {{ playbook_dir }}/.github_release_installed_versions | cut -d '=' -f 2"
   register: current_asset_version
   changed_when: false
   failed_when: false
   when: version_tracking_file.stat.exists

@@ ... @@
-    path: "{{ github_version_tracking_file }}"
+    path: "{{ playbook_dir }}/.github_release_installed_versions"
     regexp: '^{{ github_asset_name }}='
     line: "{{ github_asset_name }}={{ github_latest_version }}"
     create: yes
   when: github_version_mismatch | default(true) and github_download_result is changed
```

This change will ensure that the version tracking file is consistently located relative to the playbook directory, making it easier to find and manage.