aider: Added vm-setup/sweetums.yml to .aiderignore.
# Aider chat conversation:

USER: Add a comment to sweetums.yml stating that we should be ignoring this file.

ASSISTANT: ```diff
--- vm-setup/sweetums.yml
+++ vm-setup/sweetums.yml
@@ ... @@
 ---
+# This file should be ignored
 - hosts: localhost
   become: yes
```

USER: Cmd('git') failed due to: exit code(1)
  cmdline: git commit -m Added vm-setup/sweetums.yml. --no-verify -- /home/harleypig/projects/dotfiles/vm-setup/sweetums.yml
  stdout: 'On branch master
Your branch is up to date with 'origin/master'.

Changes not staged for commit:
  (use "git add <file>..." to update what will be committed)
  (use "git restore <file>..." to discard changes in working directory)
	modified:   .aiderignore

no changes added to commit (use "git add" and/or "git commit -a")'

ASSISTANT: It seems like you're encountering an issue with a Git command outside of the scope of our current interaction. The error message indicates that there are changes in the `.aiderignore` file that have not been staged for commit. To resolve this issue, you would typically stage the changes with `git add .aiderignore` and then commit them with `git commit -m "Your commit message"`.

However, since I am here to assist with code changes and not to execute Git commands, please let me know if there's anything else you'd like to do with the `sweetums.yml` file or any other file that you have made read-write in this chat.

USER: I added these *read-write* files: .aiderignore

ASSISTANT: To address the Git error you encountered, you can add an entry to the `.aiderignore` file to ignore the `vm-setup/sweetums.yml` file. Here is the diff for the change to `.aiderignore`:

```diff
--- .aiderignore
+++ .aiderignore
@@ ... @@
 archive/*
 bin/*
 lib/*
 shell_startup.d/*
 tests/*
 vm-setup/*
+vm-setup/sweetums.yml
 vm-setup/credentials.yml
```

After applying this change, the `vm-setup/sweetums.yml` file will be ignored as per your request. Remember to stage and commit this change in your local Git repository.