aider: Create a variable for the version tracking file name and update the main tasks to read the current version from this file.
# Aider chat conversation:

USER: In the main tasks file, the first thing that should be done is get the current version from the version tracking file. The name for the version tracking file should created in the defaults main.yml so it can be overridden if needed.

ASSISTANT: First, let's create a new variable in the `defaults/main.yml` file for the version tracking file name.

```diff
--- vm-setup/roles/github_release/defaults/main.yml
+++ vm-setup/roles/github_release/defaults/main.yml
@@ ... @@
+github_version_tracking_file: "{{ playbook_dir }}/.github_release_installed_versions"
```

Next, we need to modify the `main.yml` file to ensure that the first task is to get the current version from the version tracking file, using the new variable.

```diff
--- vm-setup/roles/github_release/tasks/main.yml
+++ vm-setup/roles/github_release/tasks/main.yml
@@ ... @@
-# This role downloads a GitHub release
+# This role manages the download and version tracking of a GitHub release

+- name: Check if version tracking file exists
+  stat:
+    path: "{{ github_version_tracking_file }}"
+  register: version_tracking_file
+
+- name: Read current version of the asset from the version tracking file
+  command: "grep '{{ github_asset_name }}' {{ github_version_tracking_file }} | cut -d '=' -f 2"
+  register: current_asset_version
+  changed_when: false
+  failed_when: false
+  when: version_tracking_file.stat.exists
+
 - name: Get the latest release of {{ github_repo }} from GitHub
   uri:
     url: "https://api.github.com/repos/{{ github_repo }}/releases/latest"
@@ ... @@
-  - name: Check if version tracking file exists
-    stat:
-      path: "{{ playbook_dir }}/.github_release_installed_versions"
-    register: version_tracking_file
-
-  - name: Read current version of the asset from the version tracking file
-    command: "grep '{{ github_asset_name }}' {{ playbook_dir }}/.github_release_installed_versions | cut -d '=' -f 2"
-    register: current_asset_version
-    changed_when: false
-    failed_when: false
-    when: version_tracking_file.stat.exists
```

These changes will ensure that the version tracking file's path can be configured and that the first task in `main.yml` is to check and read the current version from this file.